Fix issue with setting Content-Type on S3 upload. Set order of clinical insights sections. Update package versions
