+++ package.json
@@ -13,28 +13,28 @@
     "dependencies": {
         "@aws-amplify/ui": "^5.9.0",
         "@aws-amplify/ui-react": "^5.3.3",
-        "@aws-sdk/client-comprehendmedical": "^3.514.0",
-        "@aws-sdk/client-polly": "^3.514.0",
-        "@aws-sdk/client-s3": "^3.514.0",
-        "@aws-sdk/lib-storage": "^3.514.0",
-        "@aws-sdk/s3-request-presigner": "^3.514.0",
-        "@cloudscape-design/collection-hooks": "^1.0.36",
-        "@cloudscape-design/components": "^3.0.529",
+        "@aws-sdk/client-comprehendmedical": "^3.529.1",
+        "@aws-sdk/client-polly": "^3.529.1",
+        "@aws-sdk/client-s3": "^3.529.1",
+        "@aws-sdk/lib-storage": "^3.529.1",
+        "@aws-sdk/s3-request-presigner": "^3.529.1",
+        "@cloudscape-design/collection-hooks": "^1.0.38",
+        "@cloudscape-design/components": "^3.0.566",
         "@cloudscape-design/design-tokens": "^3.0.34",
-        "@cloudscape-design/global-styles": "^1.0.23",
-        "ace-builds": "^1.32.6",
-        "aws-amplify": "^5.3.15",
+        "@cloudscape-design/global-styles": "^1.0.24",
+        "ace-builds": "^1.32.7",
+        "aws-amplify": "^5.3.17",
         "axios": "^1.6.7",
         "crunker": "^2.4.0",
         "dayjs": "^1.11.10",
-        "framer-motion": "^11.0.5",
+        "framer-motion": "^11.0.8",
         "lodash": "^4.17.21",
         "react": "^18.2.0",
         "react-ace": "^10.1.0",
         "react-dom": "^18.2.0",
         "react-dropzone": "^14.2.3",
         "react-hot-toast": "^2.4.1",
-        "react-router-dom": "^6.22.0",
+        "react-router-dom": "^6.22.3",
         "timekeeper": "^2.3.1",
         "use-debounce": "^10.0.0",
         "uuid4": "^2.0.3",
@@ -43,24 +43,25 @@
     "devDependencies": {
         "@trivago/prettier-plugin-sort-imports": "^4.3.0",
         "@types/lodash": "^4.14.202",
-        "@types/react": "^18.2.55",
-        "@types/react-dom": "^18.2.19",
+        "@types/react": "^18.2.64",
+        "@types/react-dom": "^18.2.21",
         "@types/uuid4": "^2.0.3",
         "@types/wavesurfer.js": "^6.0.12",
-        "@typescript-eslint/eslint-plugin": "^7.0.1",
-        "@typescript-eslint/parser": "^7.0.1",
+        "@typescript-eslint/eslint-plugin": "^7.1.1",
+        "@typescript-eslint/parser": "^7.1.1",
         "@vitejs/plugin-react": "^4.2.1",
-        "eslint": "^8.56.0",
+        "eslint": "^8.57.0",
         "eslint-config-prettier": "^9.1.0",
         "eslint-plugin-import": "^2.29.1",
         "eslint-plugin-prettier": "^5.1.3",
-        "eslint-plugin-react": "^7.33.2",
+        "eslint-plugin-react": "^7.34.0",
         "pre-commit": "^1.2.2",
         "prettier": "^3.2.5",
-        "typescript": "^5.3.3",
-        "vite": "^5.1.2",
+        "typescript": "^5.4.2",
+        "vite": "^5.1.5",
         "vite-plugin-static-copy": "^1.0.1"
     },
     "pre-commit": "lint",
-    "license": "MIT"
+    "license": "MIT",
+    "packageManager": "yarn@1.22.19+sha1.4ba7fc5c6e704fce2066ecbfb0b0d8976fe62447"
 }
