+++ src/components/Conversation/RightPanel/RightPanel.tsx
@@ -9,14 +9,12 @@ import ScrollingContainer from '@/components/Conversation/Common/ScrollingContai
 import {
     IAuraClinicalDocOutput,
     IAuraClinicalDocOutputSection,
-    IAuraClinicalDocOutputSectionNew,
     IAuraTranscriptOutput,
     ITranscriptSegments,
 } from '@/types/HealthScribe';
 
 import { HighlightId } from '../types';
 import SummarizedConcepts from './SummarizedConcepts';
-import SummarizedConceptsNew from './SummarizedConceptsNew';
 
 type RightPanelProps = {
     jobLoading: boolean;
@@ -42,38 +40,18 @@ export default function RightPanel({
         }, {});
     }, [transcriptFile]);
 
-    // Changes in November 2023 - ClinicalDocumentation.Sections contains an array of SectionName and Summary
-    // The previous style had "SUBJECTIVE" and "ASSESSMENT_AND_PLAN" subsections
-    // Return true if clinicalDocument uses the new format
-    const isNewSummarizationFormation = useMemo(() => {
-        return (
-            typeof clinicalDocument?.ClinicalDocumentation?.Sections?.[0] !== 'undefined' &&
-            'Summary' in clinicalDocument.ClinicalDocumentation.Sections[0]
-        );
-    }, [clinicalDocument]);
-
     if (jobLoading || clinicalDocument == null) {
         return <LoadingContainer containerTitle="Insights" text="Loading Insights" />;
     } else {
         return (
             <ScrollingContainer containerTitle="Insights">
-                {isNewSummarizationFormation ? (
-                    <SummarizedConceptsNew
-                        sections={clinicalDocument.ClinicalDocumentation.Sections as IAuraClinicalDocOutputSectionNew[]}
-                        highlightId={highlightId}
-                        setHighlightId={setHighlightId}
-                        segmentById={segmentById}
-                        wavesurfer={wavesurfer}
-                    />
-                ) : (
-                    <SummarizedConcepts
-                        sections={clinicalDocument.ClinicalDocumentation.Sections as IAuraClinicalDocOutputSection[]}
-                        highlightId={highlightId}
-                        setHighlightId={setHighlightId}
-                        segmentById={segmentById}
-                        wavesurfer={wavesurfer}
-                    />
-                )}
+                <SummarizedConcepts
+                    sections={clinicalDocument.ClinicalDocumentation.Sections as IAuraClinicalDocOutputSection[]}
+                    highlightId={highlightId}
+                    setHighlightId={setHighlightId}
+                    segmentById={segmentById}
+                    wavesurfer={wavesurfer}
+                />
             </ScrollingContainer>
         );
     }
