+++ src/components/Conversation/RightPanel/sectionOrder.ts
@@ -1,0 +1,9 @@
+export const SECTION_ORDER = [
+    'CHEF_COMPLIANT',
+    'HISTORY_OF_PRESENT_ILLNESS',
+    'REVIEW_OF_SYSTEMS',
+    'PAST_MEDICAL_HISTORY',
+    'PHYSICAL_EXAMINATION',
+    'ASSESSMENT',
+    'PLAN',
+];
